# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Business Platform API",
        "description": "Authentication & RBAC endpoints for the AI Business Platform. See ARCHITECTURE.md, API.md, and BACKEND.md in the repository root for the full system design.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "List employees — full directory for Owner/Admin/HR, department-scoped for managers",
                "operationId": "ebc786767076450e76aaad14eecebc05",
                "parameters": [
                    {
                        "name": "department_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "employment_status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "manager_employee_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Matches name, email, or employee number",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "first_name",
                                "last_name",
                                "hire_date",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employees returned"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create an employee (system-generates the employee number). Requires Admin, HR, or Owner.",
                "operationId": "832921d380d564a74f783afee77f78fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "hire_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "department_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "position_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "manager_employee_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "employment_type": {
                                        "type": "string",
                                        "enum": [
                                            "full_time",
                                            "part_time",
                                            "contractor",
                                            "intern"
                                        ]
                                    },
                                    "employment_status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "on_leave",
                                            "suspended",
                                            "terminated"
                                        ]
                                    },
                                    "hire_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "address": {
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "emergency_contact": {
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "bio": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "403": {
                        "description": "Missing employees.manage permission"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/employees/me": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get the authenticated user's own employee profile",
                "operationId": "aa2b47dad58cc8e00d4835e0d6f0bfdf",
                "responses": {
                    "200": {
                        "description": "Profile returned"
                    },
                    "404": {
                        "description": "No employee record is linked to this account"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/employees/{employee}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get a single employee",
                "operationId": "d0b012a682df5d590fb41b7ee73a6676",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee returned"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Archive (soft-delete) an employee",
                "operationId": "d37a3c2019f07c2fa16aa57590dbeb53",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee archived"
                    },
                    "403": {
                        "description": "Missing employees.manage permission"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update an employee. Employees may update their own contact/profile fields; only Admin/HR/Owner may change department, position, manager, status, or dates.",
                "operationId": "88adc52a51df3bb742f561a5c896b59d",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee updated"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "403": {
                        "description": "Not authorized to change one or more of the submitted fields"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/employees/{employee}/avatar": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Upload a profile picture",
                "operationId": "bc7b695ac843369fd7b6c3031fa3f9ae",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar uploaded"
                    },
                    "422": {
                        "description": "Validation failed (not an image, or too large)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Send the token returned by /api/v1/auth/register or /api/v1/auth/login as 'Authorization: Bearer {token}'.",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}